// resources/js/api/generated/models/getApiV1ServersServeridStacksStacknameVolumes200VolumesItem.ts
/**
 * Generated by orval v8.0.2 🍺
 * Do not edit manually.
 * Berth API
 * Berth: Opinionated docker compose stack management API
 * OpenAPI spec version: 1.0.0
 */
import type { GetApiV1ServersServeridStacksStacknameVolumes200VolumesItemDriverOpts } from './getApiV1ServersServeridStacksStacknameVolumes200VolumesItemDriverOpts';
import type { GetApiV1ServersServeridStacksStacknameVolumes200VolumesItemLabels } from './getApiV1ServersServeridStacksStacknameVolumes200VolumesItemLabels';
import type { GetApiV1ServersServeridStacksStacknameVolumes200VolumesItemUsedByItem } from './getApiV1ServersServeridStacksStacknameVolumes200VolumesItemUsedByItem';

export type GetApiV1ServersServeridStacksStacknameVolumes200VolumesItem = {
  created?: string;
  driver?: string;
  driver_opts?: GetApiV1ServersServeridStacksStacknameVolumes200VolumesItemDriverOpts;
  exists: boolean;
  external?: boolean;
  labels?: GetApiV1ServersServeridStacksStacknameVolumes200VolumesItemLabels;
  mountpoint?: string;
  name: string;
  scope?: string;
  used_by?: GetApiV1ServersServeridStacksStacknameVolumes200VolumesItemUsedByItem[];
};